+++ docs/conf.py
@@ -39,6 +39,7 @@ extensions = [
     "sphinx_issues",
     "sphinxext.opengraph",
     "myst_parser",
+    "click_extra.sphinx",
 ]
 
 myst_enable_extensions = [
@@ -88,77 +89,3 @@ html_last_updated_fmt = "%Y-%m-%d"
 copyright = f"{author} and contributors"
 html_show_copyright = True
 html_show_sphinx = False
-
-
-pygments_style = "ansi-click-extra-furo-style"
-
-
-from sphinx.highlighting import PygmentsBridge
-
-from click_extra.pygments import AnsiHtmlFormatter
-
-PygmentsBridge.html_formatter = AnsiHtmlFormatter  # type: ignore
-
-
-####################################
-#  pallets_sphinx_themes Patch #1  #
-####################################
-
-# XXX Compatibility workaround because of https://github.com/pallets/pallets-sphinx-themes/blob/7b69241f1fde3cc3849f513a9dd83fa8a2f36603/src/pallets_sphinx_themes/themes/click/domain.py#L9
-# Source:
-# https://github.com/pallets/click/blob/dc918b48fb9006be683a684b42cc7496ad649b83/docs/conf.py#L6-L7
-import click
-
-setattr(click._compat, "text_type", str)
-
-# Add support for ".. click:example::" and ".. click:run::" directives.
-from pallets_sphinx_themes.themes.click import domain
-
-append_orig = domain.ViewList.append
-
-
-def patched_append(*args, **kwargs):
-    """Replace the code block produced by ``.. click:run::`` directive with an ANSI
-    Shell Session (``.. code-block:: ansi-shell-session``).
-
-    Targets:
-    - [``.. sourcecode:: text`` for `Pallets-Sphinx-Themes <= 2.0.2`](https://github.com/pallets/pallets-sphinx-themes/blob/7b69241f1fde3cc3849f513a9dd83fa8a2f36603/src/pallets_sphinx_themes/themes/click/domain.py#L245)
-    - [``.. sourcecode:: shell-session`` for `Pallets-Sphinx-Themes > 2.0.2`](https://github.com/pallets/pallets-sphinx-themes/pull/62)
-    """
-    default_run_blocks = (
-        ".. sourcecode:: text",
-        ".. sourcecode:: shell-session",
-    )
-    for run_block in default_run_blocks:
-        if run_block in args:
-            args = list(args)
-            index = args.index(run_block)
-            args[index] = ".. code-block:: ansi-shell-session"
-
-    return append_orig(*args, **kwargs)
-
-
-domain.ViewList.append = patched_append
-
-
-####################################
-#  pallets_sphinx_themes Patch #2  #
-####################################
-
-# Replace the call to default ``CliRunner.invoke`` with a call to click_extra own version which is sensible to contextual color settings
-# and output unfiltered ANSI codes.
-# Fixes: <insert upstream bug report here>
-from click_extra.tests.conftest import ExtraCliRunner
-
-# Force color rendering in ``invoke`` calls.
-ExtraCliRunner.force_color = True
-
-# Brutal, but effective.
-# Alternative patching methods: https://stackoverflow.com/a/38928265
-domain.ExampleRunner.__bases__ = (ExtraCliRunner,)
-
-
-####################################
-#  Register new click directives   #
-####################################
-extensions.append("pallets_sphinx_themes.themes.click.domain")