+++ changelog.md
@@ -6,6 +6,9 @@
 This version is not released yet and is under active development.
 ```
 
+- Streamline setup of Sphinx extensions.
+- Document ``.. click:example::`` and ``.. click:run::`` Sphinx extensions.
+
 ## {gh}`3.3.4 (2022-11-14) <compare/v3.3.3...v3.3.4>`
 
 - Fix some types.